Added bench_vs_lfs2_many_open
For open+read results, and renamed bench_vs_lfs2_many ->
bench_vs_lfs2_many_creat.

Though this is raising some questions about if we're approaching random
file writes correctly... We're not really capturing growth like we are
in bench_vs_lfs2_many_open.

---

It also raised a question if amortized results make sense when each
measurement is more-or-less isolated, but in bench_vs_lfs2_many_open's
case, it does!

The mtree takes different shapes as it grows, and it's interesting to
see both the averaged result per-topology and amortized result across
all topologies.

Though I will admit it is a bit abstract and tricky to think about...