Added incremental rewriting to bench_fwrite
And adopted incremental rewriting for randome fwrite results in
bench_fwrite and bench_fwrite_tune.

The difference between the rewriting benchmarks and incremental
rewriting benchmarks is that rewriting operates on a fixed size file,
whereas incremental rewriting operates on an n-sized file where n is
the current iteration.

Full rewriting is probably the more intuitive/first approach, but it
doesn't capture how rewriting interacts with different file sizes.
Incremental rewriting, on the other hand, captures how the topology of
the file affects random writes, information that is missing from a
simple/naive random rewrite that is trying to mimic normal use.