Fixed bench_rbyd running out of space
We use a lot more space for benchmarks than for testing, so this was
hitting OOS issues.

The reason we're using a fixed size, is to avoid emubd freaking out when
block_size is _huge_. A side-effect that happens when setting DISK_SIZE
to very large values for fwrite benchmarking.