Added bench_vs_lfs2_logging to explicitly benchmark logging
Logging seems to be one of the most common use cases for littlefs, so
it's useful to have an explicit benchmark over this use case.

This also lets us show off lfsr_file_fruncate, which was added to help
make logging easier in littlefs.

Though, in practice, most of the _measurable_ improvements (API aside)
come from the improved performance of sync, and ability to reuse
erased-state even when the file is not aligned to a prog boundary.